requirements.txt
numpy
matplotlib
pandas
shapely
tqdm
lightning
scikit-learn
pyyaml
torchvision
torchaudio
basemap
xarray
zarr
kagglehub
huggingface_hub
geotiff
cartopy
geopy
neuraloperator
einops
wandb
kaggle
gcsfs
fsspec
cfgrib
cdsapi
xesmf
torch
optuna
dotenv
rasterio
geopandas
pytest
geojson
requests_cache
earthengine-api
pygeoboundaries_geolab @ git+https://github.com/N-Masi/pygeoboundaries.git
maskedtensor-nightly
pytorch_lightning
rich
hydra-core
